+++ resdk/scripts/__init__.py
@@ -1,3 +1,4 @@
 """Command line scripts."""
 from .reads import upload_reads
 from .sequp import sequp
+from .upload_and_annotate import upload_and_annotate